+++ src/utils/nameManager.ts
@@ -1,0 +1,19 @@
+/**
+ * 插件名称
+ */
+export const globalName = 'Godot Comment Align'
+
+/**
+ * logger channel 名称
+ */
+export const channelName = globalName
+
+/**
+ * section
+ */
+export const configSection = 'godotCommentAlign'
+
+/**
+ * 命令名称
+ */
+export const commandName = `godot-comment-align.formatAndAlign`